Added power sources to OpenDC. In the current form each Cluster has a single power source that is connected to all hosts in that cluster
