+++ opendc-compute/opendc-compute-topology/src/main/kotlin/org/opendc/compute/topology/specs/ClusterSpec.kt
@@ -1,0 +1,7 @@
+package org.opendc.compute.topology.specs
+
+public data class ClusterSpec (
+    val name: String,
+    val hostSpecs: List<HostSpec>,
+    val powerSource: PowerSourceSpec
+)
