+++ opendc-compute/opendc-compute-topology/src/main/kotlin/org/opendc/compute/topology/specs/PowerSourceSpec.kt
@@ -1,0 +1,11 @@
+package org.opendc.compute.topology.specs
+
+import java.util.UUID
+
+// TODO: add name to class
+public data class PowerSourceSpec(
+    val uid: UUID,
+    val name: String = "unknown",
+    val meta: Map<String, Any> = emptyMap(),
+    val totalPower: Long
+)
