+++ opendc-simulator/opendc-simulator-compute/src/main/java/org/opendc/simulator/compute/power/SimPowerSource.java
@@ -22,7 +22,6 @@
 
 package org.opendc.simulator.compute.power;
 
-import java.time.InstantSource;
 import org.opendc.simulator.compute.cpu.SimCpu;
 import org.opendc.simulator.engine.FlowEdge;
 import org.opendc.simulator.engine.FlowGraph;
@@ -33,15 +32,13 @@ import org.opendc.simulator.engine.FlowSupplier;
  * A {@link SimPsu} implementation that estimates the power consumption based on CPU usage.
  */
 public final class SimPowerSource extends FlowNode implements FlowSupplier {
-    private final InstantSource clock;
-
     private long lastUpdate;
 
     private double powerDemand = 0.0f;
     private double powerSupplied = 0.0f;
     private double totalEnergyUsage = 0.0f;
 
-    private FlowEdge cpuEdge;
+    private FlowEdge muxEdge;
 
     private double capacity = Long.MAX_VALUE;
 
@@ -55,7 +52,7 @@ public final class SimPowerSource extends FlowNode implements FlowSupplier {
      * @return <code>true</code> if the InPort is connected to an OutPort, <code>false</code> otherwise.
      */
     public boolean isConnected() {
-        return cpuEdge != null;
+        return muxEdge != null;
     }
 
     /**
@@ -94,9 +91,7 @@ public final class SimPowerSource extends FlowNode implements FlowSupplier {
     public SimPowerSource(FlowGraph graph) {
         super(graph);
 
-        this.clock = graph.getEngine().getClock();
-
-        lastUpdate = graph.getEngine().getClock().millis();
+        lastUpdate = this.clock.millis();
     }
 
     ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
@@ -109,7 +104,7 @@ public final class SimPowerSource extends FlowNode implements FlowSupplier {
         double powerSupply = this.powerDemand;
 
         if (powerSupply != this.powerSupplied) {
-            this.pushSupply(this.cpuEdge, powerSupply);
+            this.pushSupply(this.muxEdge, powerSupply);
         }
 
         return Long.MAX_VALUE;
@@ -159,11 +154,11 @@ public final class SimPowerSource extends FlowNode implements FlowSupplier {
 
     @Override
     public void addConsumerEdge(FlowEdge consumerEdge) {
-        this.cpuEdge = consumerEdge;
+        this.muxEdge = consumerEdge;
     }
 
     @Override
     public void removeConsumerEdge(FlowEdge consumerEdge) {
-        this.cpuEdge = null;
+        this.muxEdge = null;
     }
 }
